Feature added to define a base element and only collect content from its children, defaults to `body`.
Currently the base element is the `body` tag but it can be useful to specify a different element to use as a base for content conversion.  To select an element as the base element, add the baseElement key to your options object and set it to the tag of the element you want to describe.  This feature supports filtering on css selectors to help with specifying the correct content.  If the routine can't find the specified element it now defaults to not returning any content.